data-ingestion/processing/s3-lambda/requirements.txt
# Lambda dependencies
boto3>=1.28.0
psycopg2-binary>=2.9.9
pyarrow==15.0.0  # For Parquet support
